Changed Market to load on demand
Changed to lazy load the Market via knockout.
A lot of markup was changed for this and needs to be more fully tested.
Items not being loaded in yet, those will need to be loaded on demand as
well.
Also changed Entity Framework to turn off lazy loading in order to
serialize some of the entities that are being sent as JSON.