update entities from DB with FKs
Well, kinda had to scrap and redo the entities again. Although it looks
like git is just marking them as changes, which is nice. But now all the
FKs are present on the entities, so that's nice. Also changed the js to
use the relative URL after adding that to the EVEFIRSTMATE global
object.